implementado leitura de parametros scales
