+++ ccode.depend
@@ -1,12 +1,10 @@
 # depslib dependency file v1.0
-1634259742 source:c:\users\user\desktop\projeto_tcc\crq_network\ccode\main.c
+1634328819 source:c:\users\user\desktop\projeto_tcc\crq_network\ccode\main.c
 	<stdio.h>
 	<stdlib.h>
 	<math.h>
 	<assert.h>
-	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image.h"
-	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image_write.h"
-	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image_resize.h"
+	"cnn_lib.h"
 
 1594640766 c:\workspace\vcpkg\installed\x64-windows\include\stb_image.h
 	"stb_image.h"
@@ -41,4 +39,13 @@
 	<string.h>
 	<math.h>
 	<stdlib.h>
+
+1634330855 c:\users\user\desktop\projeto_tcc\crq_network\ccode\cnn_lib.h
+	<stdio.h>
+	<stdlib.h>
+	<math.h>
+	<assert.h>
+	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image.h"
+	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image_write.h"
+	"C:\workspace\vcpkg\installed\x64-windows\include\stb_image_resize.h"
